requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
apscheduler==3.10.4
sqlalchemy==2.0.35
anthropic==0.34.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
httpx==0.27.2
jinja2==3.1.4
python-dotenv==1.0.1
pydantic-settings==2.5.2
